Tested and documented local_conn
