+++ core/connection.py
@@ -18,7 +18,6 @@ Classes: Connections
 """
 from abc import ABC, abstractmethod
 import logging
-from configparser import NoOptionError
 from core.utils import set_log_level
 
 class Connection(ABC):
@@ -38,6 +37,7 @@ class Connection(ABC):
         self.log = logging.getLogger(type(self).__name__)
         self.msg_processor = msg_processor
         self.params = params
+        self.registered = {}
         set_log_level(params, self.log)
 
     @abstractmethod
@@ -53,73 +53,5 @@ class Connection(ABC):
         """Set up the passed in handler to be called for any message on the
         destination.
         """
-
-class LocalConnection(Connection):
-    """A special connection that can link Sensors and Actuators to other
-    Actuators. One of three optional parameters can be provided:
-    "OnEq": if message equal to this parameter is received publish "ON", works
-    with String messages
-    "OnGT": if the message is greater than this value, publsih "ON", only works
-    with messages that can be parsed to float.
-    "OnLT": if the message is less than this value, publish "ON", only works
-    with messages that can be parsed to float.
-
-    If none of the optional parameters are present, the message is passes as is.
-    All messages that don't match the comparison results in "OFF". If more than
-    one is defined, OnEQ is first and OnGT is second and OnLT is last.
-
-    This can be used to, for example, turn on an LED attached to a GPIO pin
-    configured with RpiGpioActuator when a sensor has a given value.
-    """
-
-    def __init__(self, msg_processor, params):
-        super().__init__(msg_processor, params)
-
-        self.registered = {}
-        self.eq = None
-        self.gt = None
-        self.lt = None
-
-        try:
-            self.eq = params("OnEq")
-        except NoOptionError:
-            pass
-        try:
-            if not self.eq:
-                self.gt = float(params("OnGT"))
-        except NoOptionError:
-            pass
-        try:
-            if not self.eq and not self.gt:
-                self.lt = float(params("OnLT"))
-        except NoOptionError:
-            pass
-
-    def publish(self, message, destination):
-        """Send the message or, if defined, translate the message to ON or OFF."""
-        if destination in self.registered:
-            try:
-                send = message
-                if self.eq and message == self.eq:
-                    send = "ON"
-                elif self.eq:
-                    send = "OFF"
-                elif self.gt and float(message) > self.gt:
-                    send = "ON"
-                elif self.gt:
-                    send = "OFF"
-                elif self.lt and float(message) < self.lt:
-                    send = "ON"
-                elif self.lt:
-                    send = "OFF"
-                self.log.info("Received message %s, forwarding %s to %s", message,
-                              send, destination)
-                self.registered[destination](send)
-            except ValueError:
-                self.log.error("'%s' cannot be parsed to float!", message)
-        else:
-            self.log.error("There is no handler registered for %s", destination)
-
-    def register(self, destination, handler):
         self.log.info("Registering destination %s", destination)
         self.registered[destination] = handler
