+++ test.ini
@@ -1,55 +1,3 @@
-[Sensor0]
-Class = heartbeat.heartbeat.Heartbeat
-Connection = MQTT
-Poll = 60
-Num-Dest = heartbeat/num
-Str-Dest = heartbeat/str
-Level = INFO
-
-[Sensor1]
-class = exec.exec_sensor.ExecSensor
-Connection = MQTT
-Poll = 10
-Script = echo Exec hello from sensor 1
-Destination = sensor1/hello
-Level = INFO
-
-[Sensor2]
-Class = heartbeat.heartbeat.Heartbeat
-Connection = openHAB
-Poll = 60
-Num-Dest = TestHB_Num
-Str-Dest = TestHB_Str
-Level = INFO
-
-[Actuator0]
-Class = exec.exec_actuator.ExecActuator
-Connection = MQTT
-Command = /usr/bin/curl -d '' http://rokuultra.koshak.net:8060/keypress/home
-CommandSrc = roku/home
-ResultsDest = roku/results
-Timeout = 10
-Level=INFO
-
-;[Actuator0]
-;Class = exec.exec_actuator.ExecActuator
-;Connection =
-;Command = echo
-;CommandSrc = actuator0/hello
-;ResultsDest = actuator0/hello/results
-;Timeout = 10
-;Level = INFO
-
-[Actuator1]
-Class = exec.exec_actuator.ExecActuator
-Connection = openHAB
-Command = echo
-CommandSrc = Test_Act1
-ResultsDest = Test_Act1_Results
-Timeout = 10
-Level = INFO
-
-
 [Logging]
 File = /tmp/sensorReporter.log
 MaxSize = 67108864
@@ -68,11 +16,67 @@ Port = 1883
 Keepalive = 10
 RootTopic = sensor_reporter/test
 TLS = NO
-Level = DEBUG
+Level = INFO
 
 [Connection1]
 Class = openhab_rest.rest_conn.OpenhabREST
 Name = openHAB
 URL = http://10.10.1.127:8080
 RefreshItem = Test_Refresh
-Level = DEBUG
+Level = INFO
+
+[Connection2]
+Class = local.local_conn.LocalConnection
+Name = Local
+OnLt = 20000
+Level = INFO
+
+[Sensor0]
+Class = heartbeat.heartbeat.Heartbeat
+Connection = MQTT
+Poll = 60
+Num-Dest = heartbeat/num
+Str-Dest = heartbeat/str
+Level = INFO
+
+[Sensor1]
+class = exec.exec_sensor.ExecSensor
+Connection = MQTT
+Poll = 10
+Script = echo Exec hello from sensor 1
+Destination = sensor1/hello
+Level = INFO
+
+[Sensor2]
+Class = heartbeat.heartbeat.Heartbeat
+Connection = openHAB,Local
+Poll = 5
+Num-Dest = TestHB_Num
+Str-Dest = TestHB_Str
+Level = INFO
+
+[Actuator0]
+Class = exec.exec_actuator.ExecActuator
+Connection = MQTT
+Command = /usr/bin/curl -d '' http://rokuultra.koshak.net:8060/keypress/home
+CommandSrc = roku/home
+ResultsDest = roku/results
+Timeout = 10
+Level=INFO
+
+[Actuator1]
+Class = exec.exec_actuator.ExecActuator
+Connection = openHAB
+Command = echo
+CommandSrc = Test_Act1
+ResultsDest = Test_Act1_Results
+Timeout = 10
+Level = INFO
+
+[Actuator2]
+Class = exec.exec_actuator.ExecActuator
+Connection = Local
+Command = echo
+CommandSrc = TestHB_Num
+Timeout = 10
+Level = INFO
